requirements.txt
flask
python-dotenv
google-generativeai
pydantic
pillow
werkzeug